Desenvolvimento da class Digital House Manager

+++ src/ProfessorAdjunto.kt
@@ -1,4 +1,4 @@
-class ProfessorAdjunto (nome: String, sobrenome: String, codigoProfessor: Int, val qtdHorasDeMonitoria: Float) :
+class ProfessorAdjunto (nome: String, sobrenome: String, codigoProfessor: Int, val qtdHorasDeMonitoria: Int) :
     Professor(nome, sobrenome, codigoProfessor) {
 
 }
